requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.0.74
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
premailer==3.10.0
psutil==5.9.6
py-cpuinfo==9.0.0
pycryptodome==3.19.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.205
urllib3==2.0.7
yagmail==0.15.293
zipp==3.17.0
